+++ Runtime/Audio/AudioVolumeAnimation.cs
@@ -6,7 +6,7 @@ namespace ActionCode.AnimationSystem
     /// <summary>
     /// Animates the volume of an AudioSource, creating a smooth fade transition between an initial and a final value over a specified duration.
     /// </summary>
-    [AddComponentMenu("Animation/Audio/Volume")]
+    [AddComponentMenu("Animation/Audio/Volume Animation")]
     public sealed class AudioVolumeAnimation : AbstractAnimation
     {
         [Tooltip("The local AudioSource component.")]