+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Animation ResetSpeed function
+- AnimationSequence PlayAsync function
+
+### Changed
+- Make serialized fields public
 
 ## [0.4.0] - 2026-07-14
 ### Added